[refactor/#42] onClick 속성 추가
